Update to Minecraft 1.19.4.
Update to Minecraft 1.19.4. Add Alpha Winter alt world preset. Add world
preset tooltips.

+++ src/main/java/mod/bespectacled/modernbeta/util/NbtReader.java
@@ -42,7 +42,7 @@ public class NbtReader {
         return NbtUtil.readDoubleOrThrow(key, this.compound);
     }
     
-    public double readDouble(String key, float alternate) {
+    public double readDouble(String key, double alternate) {
         return NbtUtil.readDouble(key, this.compound, alternate);
     }
